Simplify body of Scanner::getData
When I was reviewing this section of code, as part of a larger
code investigation, the final if block in Scanner::getData was harder to
read than I felt that it should have been. Secondly, a string was
created twice, when it would be less error-prone to create it once and
reuse it. Given these two points, I've created this commit, to simplify
the code and improve readability.